+++ src/index.ts
@@ -1,5 +1,9 @@
 export * from './git-adapter'
 export { Commit } from './model/commit'
 export { CommitDraft } from './model/commit-draft'
-export { ContentEntry, ContentEntryMetadata } from './model/content-entry'
+export {
+  ContentEntry,
+  ContentEntryData,
+  ContentEntryMetadata,
+} from './model/content-entry'
 export { ContentEntryDraft } from './model/content-entry-draft'